+++ src/pages/task/inprogressTask/IndexInprogressTask.tsx
@@ -411,7 +411,9 @@ const IndexInprogressTask = () => {
               </span>
               <br />
               <span style={{ color: color.Grey, fontSize: "12px" }}>
-                <UserOutlined style={{ padding: "0 4px 0 0" }} />
+                <UserOutlined
+                  style={{ padding: "0 4px 0 0", verticalAlign: 0.5 }}
+                />
                 {row.task_create_by}
               </span>
             </>