+++ src/pages/task/newTask/IndexNewTask.tsx
@@ -330,7 +330,9 @@ const IndexNewTask = () => {
               </span>
               <br />
               <span style={{ color: color.Grey }}>
-                <UserOutlined style={{ padding: "0 4px 0 0" }} />
+                <UserOutlined
+                  style={{ padding: "0 4px 0 0", verticalAlign: 0.5 }}
+                />
                 {row.create_by}
               </span>
             </>