+++ CHANGELOG.md
@@ -1,4 +1,12 @@
 ## [Version - 0]
+## Feature [v0.25.00] 2023-04-07
+- add promotions feature
+## Feature [v0.24.00] 2023-04-06
+- added report droner feature (PDF) 
+## Fix [v0.23.02] 2023-03-30
+- fixed condition delete series drone
+- store path logo in the DB.
+- fixed count data location price
 ## Fix [v0.23.01] 2023-03-20
 - fixed search plants
 - fixed current page